Allowed images narrower than 800 pixels
Previously, the bot would try to ensure that it didn't use an image
narrower than 800 pixels. This would sometimes cause it to download lots
of images trying to find a valid one, and eventually end up stalling
forever. To prevent this, images that are narrower than 800 pixels are
now scaled up.